speech-to-text-service/pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0

[project]
name = "speech-to-text"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "sounddevice>=0.5.1",
    "soundfile>=0.13.1",
    "nvidia-riva-client>=2.17.0",
    "workmesh",
    "pyaudio>=0.2.14",
]

[tool.uv.sources]
workmesh = { workspace = true }

[dependency-groups]
dev = [
    "typer>=0.20.0",
]
